attach form to notification & make notifications show when app is not running
